Rename file _images -> _multiscale_image
