+++ apis/python/src/tiledbsoma/__init__.py
@@ -165,7 +165,7 @@ from ._general_utilities import (
 from ._indexer import IntIndexer, tiledbsoma_build_index
 from ._measurement import Measurement
 from ._scene import Scene
-from ._images import MultiscaleImage
+from ._multiscale_image import MultiscaleImage
 from ._point_cloud import PointCloud
 from ._sparse_nd_array import SparseNDArray, SparseNDArrayRead
 from .options import SOMATileDBContext, TileDBCreateOptions, TileDBWriteOptions
