+++ apis/python/src/tiledbsoma/_factory.py
@@ -27,8 +27,8 @@ from . import (
     _dataframe,
     _dense_nd_array,
     _experiment,
-    _images,
     _measurement,
+    _multiscale_image,
     _point_cloud,
     _scene,
     _soma_object,
@@ -219,7 +219,7 @@ def _type_name_to_cls(type_name: str) -> Type[AnySOMAObject]:
             _dense_nd_array.DenseNDArray,
             _experiment.Experiment,
             _measurement.Measurement,
-            _images.MultiscaleImage,
+            _multiscale_image.MultiscaleImage,
             _sparse_nd_array.SparseNDArray,
             _scene.Scene,
             _point_cloud.PointCloud,
