+++ apis/python/src/tiledbsoma/_multiscale_image.py
